don't use thread local if we don't have to
