+++ search-client/src/commonMain/kotlin/com/jillesvangurp/ktsearch/RoundRobinNodeSelector.kt
@@ -4,12 +4,12 @@ interface IndexProvider {
     fun get(): Int
     fun set(value: Int)
 }
-expect fun indexProvider(initialIndex: Int): IndexProvider
+expect fun simpleIndexProvider(initialIndex: Int=0): IndexProvider
 
 class RoundRobinNodeSelector(
     private val nodes: Array<out Node>
 ) : NodeSelector {
-    private val index = indexProvider(0)
+    private val index = simpleIndexProvider()
     override fun selectNode(): Node {
         val result = nodes[index.get()]
         index.set((index.get() + 1).mod(nodes.size))
