--- search-client/src/jsMain/kotlin/com/jillesvangurp/ktsearch/indexProvider.kt
@@ -1,11 +1,0 @@
-package com.jillesvangurp.ktsearch
-
-class SimpleIndexProvider(private var index: Int) : IndexProvider {
-    override fun get(): Int = index
-
-    override fun set(value: Int) {
-        index = value
-    }
-}
-
-actual fun indexProvider(initialIndex: Int): IndexProvider = SimpleIndexProvider(initialIndex)
