+++ search-client/src/jsMain/kotlin/com/jillesvangurp/ktsearch/nodeselectors.kt
@@ -1,0 +1,11 @@
+package com.jillesvangurp.ktsearch
+
+actual fun simpleIndexProvider(initialIndex: Int): IndexProvider = object: IndexProvider {
+    private var index: Int = initialIndex
+
+    override fun get(): Int = index
+
+    override fun set(value: Int) {
+        index = value
+    }
+}
