+++ search-client/src/jvmMain/kotlin/com/jillesvangurp/ktsearch/nodeselectors.kt
@@ -1,0 +1,12 @@
+package com.jillesvangurp.ktsearch
+
+import java.util.concurrent.atomic.AtomicInteger
+
+actual fun simpleIndexProvider(initialIndex: Int): IndexProvider = object: IndexProvider {
+    // use AtomicInteger for thread safety
+    private val index = AtomicInteger(initialIndex)
+
+    override fun get(): Int = index.get()
+
+    override fun set(value: Int) = index.set(value)
+}
